app/backend/requirements.txt
aiohttp==3.10.11
azure-cognitiveservices-speech==1.38.0
azure-identity==1.19.0
azure-search-documents==11.6.0
azure-storage-blob==12.24.0
cffi==1.17.1
gunicorn==23.0.0
openai==1.109.1
pydantic>=2,<3
python-dotenv==1.0.1
rich==13.9.4
PyYAML>=6,<7
Jinja2>=3,<4